requirements.txt
ultralytics
matplotlib
pillow
pandas
scikit-image
opencv-python
